Revert moving focus effects
